Rename defaultStrategy file
