+++ tests/strategy/strats/strategy_test_v2.py
